+++ src/main/java/net/mcreator/element/converter/v2020_5/GUICoordinateConverter.java
@@ -17,24 +17,32 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.mcreator.element.converter.legacy;
+package net.mcreator.element.converter.v2020_5;
 
 import com.google.gson.JsonElement;
 import net.mcreator.element.GeneratableElement;
 import net.mcreator.element.converter.IConverter;
 import net.mcreator.element.parts.gui.GUIComponent;
-import net.mcreator.element.types.Overlay;
+import net.mcreator.element.parts.gui.SizedComponent;
+import net.mcreator.element.types.GUI;
 import net.mcreator.workspace.Workspace;
 
-public class OverlayCoordinateConverter implements IConverter {
+public class GUICoordinateConverter implements IConverter {
 
 	@Override
 	public GeneratableElement convert(Workspace workspace, GeneratableElement input, JsonElement jsonElementInput) {
-		Overlay gui = (Overlay) input;
+		GUI gui = (GUI) input;
+
+		gui.width = convert(gui.width);
+		gui.height = convert(gui.height);
 
 		for (GUIComponent component : gui.components) {
 			component.x = convert(component.getX());
 			component.y = convert(component.getY());
+			if (component instanceof SizedComponent) {
+				((SizedComponent) component).width = convert(((SizedComponent) component).width);
+				((SizedComponent) component).height = convert(((SizedComponent) component).height);
+			}
 		}
 
 		return gui;
